Added filter for useless GetValue calls
If the course is being launched for the very first time, SCORM requires
cmi.entry to return "ab-initio". This means there will be no
pre-existing data, such as cmi.location, suspend_data, or score. We
should therefore prevent Captivate from asking for these items, because
it will throw a pointless error in the LMS.